# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  reponse.WebResponseFailed:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  reponse.WebResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
  socialmedia.Socialmedia:
    properties:
      created_at:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - name
    - social_media_url
    type: object
  socialmedia.SocialmediaCreateRequest:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    type: object
  token.Tokens:
    properties:
      access_token:
        type: string
    type: object
  user.UserAuthenticate:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  user.UserCreateRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  user.UserCreateResponse:
    properties:
      Username:
        type: string
      age:
        type: integer
      email:
        type: string
    type: object
host: localhost:9999
info:
  contact: {}
paths:
...
  /socialmedia/createsocialmedia/:
    post:
      consumes:
      - application/json
      description: how to create social media
      parameters:
      - description: Input Data Social Media
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/socialmedia.SocialmediaCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/socialmedia.Socialmedia'
              type: object
      security:
      - Bearer: []
      summary: Create Social Media example
      tags:
      - Social Media
  /socialmedia/deleteocialmedia/{id}:
    delete:
      consumes:
      - application/json
      description: how to delete social media
      parameters:
      - description: id social media
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reponse.WebResponseSuccess'
      security:
      - Bearer: []
      summary: Delete Social Media example
      tags:
      - Social Media
  /socialmedia/getall:
    get:
      consumes:
      - application/json
      description: how to get all social media
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/socialmedia.Socialmedia'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get All Social Media example
      tags:
      - Social Media
  /socialmedia/getone/{id}:
    get:
      consumes:
      - application/json
      description: how to get social media
      parameters:
      - description: id social media
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/socialmedia.Socialmedia'
              type: object
      security:
      - Bearer: []
      summary: Get Social Media example
      tags:
      - Social Media
  /socialmedia/updatesocialmedia/{id}:
    put:
      consumes:
      - application/json
      description: how to update social media
      parameters:
      - description: id social media
        in: path
        name: id
        type: integer
      - description: Input Data User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/socialmedia.SocialmediaCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/socialmedia.Socialmedia'
              type: object
      security:
      - Bearer: []
      summary: update Social Media example
      tags:
      - Social Media
  /user/login:
    post:
      consumes:
      - application/json
      description: do login
      parameters:
      - description: Login User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserAuthenticate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/token.Tokens'
              type: object
        "400":
          description: Input tidak valid
          schema:
            $ref: '#/definitions/reponse.WebResponseFailed'
        "500":
          description: Error Server Side
          schema:
            $ref: '#/definitions/reponse.WebResponseFailed'
      summary: Login example
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      description: do register
      parameters:
      - description: Input Data User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/user.UserCreateResponse'
              type: object
        "400":
          description: Input tidak valid
          schema:
            $ref: '#/definitions/reponse.WebResponseFailed'
        "500":
          description: Error Server Side
          schema:
            $ref: '#/definitions/reponse.WebResponseFailed'
      summary: Register example
      tags:
      - user
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"